Updated to include support for daily metrics
Note: compared to v0.1.0, the format of the output has changed. It used to contain
a Metrics attribute containing a map[string]float64 with the hourly metrics. Because
open-meteo introduced daily metrics, these are now renamed to HourlyMetrics and
DailyMetrics.

Breaking backwards compatibility was a reasonable trade-off as the module was only
released 2 days ago.